+++ src/extension.ts
@@ -2,7 +2,9 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 
-import { grabOnYCoord, grabStatically, showCalibrationText } from './demo-commands';
+import {
+    grabOnYCoord, grabStatically, openCommunication, showCalibrationText
+} from './demo-commands';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -37,6 +39,7 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(grabStatically());
   context.subscriptions.push(grabOnYCoord());
   context.subscriptions.push(showCalibrationText(calibrationUri));
+  context.subscriptions.push(openCommunication());
 }
 
 // This method is called when your extension is deactivated